add parameter to control the rapidity integration range for dilepton observables
We introduce two new parameters dilepton_y_i and dilepton_y_f to specify
the rapidity range for the dilepton observables. We use a fixed
Nrapidity = 4 for now for the rapidity integration using the trapizold
method. I think such setting should be sufficient to deal with most of
the experimental requests.